requirements.txt
click==8.0.3
Click==7.0
cmsis_svd==0.4
cmsis_svd.egg==info
cxxfilt==0.3.0
ipython==8.12.3
jsonpickle==3.0.1
libclang==15.0.6.1
llvmlite==0.41.1
matplotlib==3.5.1
matplotlib==3.1.2
pandas==1.3.4
pandas_stubs==1.2.0.35
python-dotenv==1.0.1
z3_solver==4.8.12.0
